+++ src/main/java/com/projectBackend/project/crawling/crawlMetal/CrawlMetalSchedularService.java
@@ -11,6 +11,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 
 import java.util.List;
+import java.util.Optional;
 
 @Slf4j
 @Service
@@ -27,25 +28,42 @@ public class CrawlMetalSchedularService {
         ResponseEntity<String> response = restTemplate.getForEntity(url, String.class);
         String data = response.getBody();
 
-        //assert 문은 주어진 조건이 참이면 계속 진행하고, 거짓이라면 AssertionError를 발생시켜 프로그램을 중단
-        assert data != null;
-        List<CrawlMetalDto> dataDtoList = objectMapper.readValue(data, new TypeReference<List<CrawlMetalDto>>() {});
+        if (data != null) {
+            List<CrawlMetalDto> dataDtoList = objectMapper.readValue(data, new TypeReference<List<CrawlMetalDto>>() {});
 
-        // 데이터 최신화를 위해 삭제
-        crawlMetalRepository.deleteAll();
+            // 데이터 최신화를 위해 삭제
+            crawlMetalRepository.deleteAll();
 
-        for (CrawlMetalDto metalDto : dataDtoList) {
-            CrawlMetalEntity crawlMetalEntity = CrawlMetalEntity.builder()
-                    .crawlMetalName(metalDto.getName())
-                    .crawlMetalUnits(metalDto.getUnits())
-                    .crawlMetalPrice(metalDto.getPrice())
-                    .crawlMetalYesterday(metalDto.getYesterday())
-                    .crawlMetalRate(metalDto.getRate())
-                    .crawlMetalDate(metalDto.getDate())
-                    .crawlMetalExchange(metalDto.getExchange())
-                    .build();
+            for (CrawlMetalDto metalDto : dataDtoList) {
+                // 이름으로 기존 데이터 조회
+                Optional<CrawlMetalEntity> existingEntityOptional = crawlMetalRepository.findByCrawlMetalName(metalDto.getName());
 
-            crawlMetalRepository.save(crawlMetalEntity);
+                if (existingEntityOptional.isPresent()) {
+                    // 기존 데이터가 있으면 업데이트
+                    CrawlMetalEntity crawlMetalEntity = existingEntityOptional.get();
+                    crawlMetalEntity.setCrawlMetalUnits(metalDto.getUnits());
+                    crawlMetalEntity.setCrawlMetalPrice(metalDto.getPrice());
+                    crawlMetalEntity.setCrawlMetalYesterday(metalDto.getYesterday());
+                    crawlMetalEntity.setCrawlMetalRate(metalDto.getRate());
+                    crawlMetalEntity.setCrawlMetalDate(metalDto.getDate());
+                    crawlMetalEntity.setCrawlMetalExchange(metalDto.getExchange());
+
+                    crawlMetalRepository.save(crawlMetalEntity);
+                } else {
+                    // 기존 데이터가 없으면 새로 추가
+                    CrawlMetalEntity crawlMetalEntity = CrawlMetalEntity.builder()
+                            .crawlMetalName(metalDto.getName())
+                            .crawlMetalUnits(metalDto.getUnits())
+                            .crawlMetalPrice(metalDto.getPrice())
+                            .crawlMetalYesterday(metalDto.getYesterday())
+                            .crawlMetalRate(metalDto.getRate())
+                            .crawlMetalDate(metalDto.getDate())
+                            .crawlMetalExchange(metalDto.getExchange())
+                            .build();
+
+                    crawlMetalRepository.save(crawlMetalEntity);
+                }
+            }
         }
     }
 }
